Update TDOP parser, everything is an expression
For some reason I had a weird mental block about just treating the
expressions *as* expressions.

Had a brain blast in the shower, and thought to just treat this like a
small subset of a normal programming language, rather than a grammar
that's specific to some "config language".

It's not a "section" or a "named element", they're just declarations.
"Elements" are expressions. That made things substantially more
straightforward to understand.