Deadass simple query function
This may be the single most simple way I've come up with to query the
data from a series of dicts/lists.

The only problem is, you lose a *lot* of data from the table. Worse user
experience potentially, though the overhead it adds is absolutely
trivial.

Need to draft it out more, and think of the pros/cons.