Fix include objects, remove constraint number/objs
Constraints can just be stored as an array of literal paths. You can
only supply a constraint list once, at the top level of the parent file.

It would be too convoluted having a child constrain a sub-child.

Include(s) were broken, as I hadn't fully thought through the object
structure for them. It is now an array of include objects. Each of which
containts a literal path, and the target node at which to insert.